auto message - minor changes

+++ to_del.sh
@@ -1,4 +1,4 @@
-clang++ -Wall -Werror -Wextra -std=c++98 *.cpp && ./a.out && rm a.out
+clear &&clang++ -Wall -Werror -Wextra -std=c++98 *.cpp && ./a.out && rm a.out
 
 
 echo
